+++ src/lib.rs
@@ -108,6 +108,7 @@ pub mod prelude {
     pub use crate::preprocessing::text::char_ngram_vectorizer::{
         Analyzer, CharacterNGramVectorizer,
     };
+    pub use crate::preprocessing::text::count_vectorizer::{CountVectorizer, Tokenizer};
     pub use crate::preprocessing::time_since::{ReferenceTime, TimeSince, TimeUnit};
     pub use crate::preprocessing::winsorizer::Winsorizer;
     pub use crate::time_series::cyclical::CyclicalEncoder;